added no such user problem handling

+++ src/main/java/ru/itis/services/WishListService.java
@@ -21,13 +21,16 @@ public class WishListService {
     }
 
     public WishList addNewWishList(String title, String token) {
-        User currentUser = userService.findUserByToken(token).get();
-        WishList newWL = WishList.builder()
-                .title(title)
-                .author(currentUser)
-                .build();
-        currentUser.getWishLists().add(newWL);
-        return wishListRepository.save(newWL);
+        if (userService.findUserByToken(token).isPresent()) {
+            User currentUser = userService.findUserByToken(token).get();
+            WishList newWL = WishList.builder()
+                    .title(title)
+                    .author(currentUser)
+                    .build();
+            currentUser.getWishLists().add(newWL);
+            return wishListRepository.save(newWL);
+        }
+        return null;
     }
 
     public Optional<WishList> findWishListById(Long wishListId) {
@@ -37,10 +40,12 @@ public class WishListService {
     public boolean removeByTitle(String title, String token) {
         Optional<WishList> wishListCandidate = wishListRepository.findByTitle(title);
         if (wishListCandidate.isPresent()) {
-            User currentUser = userService.findUserByToken(token).get();
-            currentUser.getWishLists().remove(wishListCandidate.get());
-            userService.save(currentUser);
-            wishListRepository.delete(wishListCandidate.get());
+            if (userService.findUserByToken(token).isPresent()) {
+                User currentUser = userService.findUserByToken(token).get();
+                currentUser.getWishLists().remove(wishListCandidate.get());
+                userService.save(currentUser);
+                wishListRepository.delete(wishListCandidate.get());
+            }
             return true;
         } else {
             return false;
